Create a mainmenu class for the other functionalities
